api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
azure-identity>=1.18.0
azure-search-documents>=11.6.0
openai>=1.50.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-instrumentation-fastapi>=0.48b0
azure-monitor-opentelemetry-exporter>=1.0.0b28
